feature: added note in gitignore bugfix: camelCase into kebab-case
