+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './components/app/app';
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(
